requirements.txt
fs
paramiko
svn
scipy
driver2200087==0.4
argparse==1.2.1
wsgiref==0.1.2
matplotlib==1.4.3
jsmin==2.1.2
psycopg2==2.6.1
dataset==0.6.0
pycallgraph==1.0.1
arrow==0.6.0
qrcode==5.1
splinter==0.7.0
iec60063==0.1
Flask-User==0.6.1
beautifulsoup4==4.3.2
Flask-Principal==0.4.0
PyPDF2==1.23
Flask-Migrate==1.4.0
twine==1.5.0
Flask-Assets==0.10
lxml==3.4.1
idstring==1.0.3
wheel==0.24.0
cssmin==0.2.0
Flask==0.10.1
Jinja2==2.7.3
alembic==0.7.6
sqlalchemy_utils==0.30.16
pillow==2.9.0
Sphinx==1.3.1
